fix: NAPI surpress unthrowable exceptions
By default, throwing an exception on a terminating environment (eg.
worker threads) will cause a fatal exception, terminating the Node
process. This is to provide feedback to the user of the runtime error,
as it is impossible to pass the error to JavaScript when the environment
is terminating. In order to bypass this behavior such that the Node
process will not terminate, define the preprocessor directive
NODE_API_SWALLOW_UNTHROWABLE_EXCEPTIONS.

Taken from node-addon-api documentation page https://github.com/nodejs/node-addon-api/blob/main/doc/setup.md#installation-and-usage