Move the emit for the original error into the stopPoll block.
When the original error occurs, it is ok to emit this, all other
polls will also likely emit this error as well. It should be clean
to emit this only once, and also include it in the callback when it
is finished.

This allows non-fatal errors to retry as well since they will fall into
the `this.pollActive` block as well.