Fix startPolling to use callback vs over emitted errors. (#631)
* Fix #548 allow multiple polling to fail.

Allow multiple transfers to emit errors,
however only attempt to stop polling if
the polling is currently active.

* Ensure callback is utilized, prevent FATAL_ERROR

The way this originally worked, it was possible to
over emit errors, which can cause node to lose track
of who can catch these. This specifically can cause many
`FATAL ERROR:  LIBUSB_TRANSFER_ERROR` or sometimes stalls.

* Fix linter issue

* Add test to ensure callback is occuring

* Move done assertion in test to call back.

Assert the end emitted has 100 packets run.

Ensure startPoll always calls callback when finished.

* Actually migrate the callback into the end case

* Move the emit for the original error into the stopPoll block.

When the original error occurs, it is ok to emit this, all other
polls will also likely emit this error as well. It should be clean
to emit this only once, and also include it in the callback when it
is finished.

This allows non-fatal errors to retry as well since they will fall into
the `this.pollActive` block as well.

* Update tsc/usb/endpoint.ts

* Fix usb.coffee test assertion

* Change the callback to use a cancelled boolean

* Fix lint

* Fix tests that run on test device

---------

Co-authored-by: Rob Moran <github@thegecko.org>