Nullchecks and errorchecks in descriptor parsing
There were a couple places in device.cc where libusb could fail or could
return nulls that weren't checked; this PR adds checks there.

Specifically,
- libusb_get_config_descriptor() is nonblocking but it allocates and
checks that allocation, so we need to check its return value
- the ->extra fields in config, interface, and endpoint descriptors,
which gather unparsed data, can sometimes be returned as NULL. The
->extra_length is always 0 in such cases, but calling memmove(dest,
NULL, 0) as happens inside the eventual napi buffer copies we do on the
extras is formally UB and practically leads to crashes.

Closes #762